Move submit action to verify view
